refactor: Rename eureka-server to registry
