+++ registry/.gitignore
