+++ registry/src/main/java/com/github/itsAkshayDubey/eventdrivenarchitecture/registry/Registry.java
@@ -1,4 +1,4 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.eurekaserver;
+package com.github.itsAkshayDubey.eventdrivenarchitecture.registry;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,10 +6,10 @@ import org.springframework.cloud.netflix.eureka.server.EnableEurekaServer;
 
 @SpringBootApplication
 @EnableEurekaServer
-public class EurekaServerApplication {
+public class Registry {
 
 	public static void main(String[] args) {
-		SpringApplication.run(EurekaServerApplication.class, args);
+		SpringApplication.run(Registry.class, args);
 	}
 
 }
