+++ registry/src/main/resources/application.properties
@@ -3,4 +3,4 @@ eureka.client.fetch-registry=false
 eureka.instance.prefer-ip-address=true
 eureka.client.service-url.defaultZone=http://localhost:8761/eureka
 server.port=8761
-spring.application.name=eureka-server
+spring.application.name=registry
