+++ registry/src/test/java/com/github/itsAkshayDubey/eventdrivenarchitecture/registry/RegistryTests.java
@@ -1,10 +1,10 @@
-package com.github.itsAkshayDubey.eventdrivenarchitecture.eurekaserver;
+package com.github.itsAkshayDubey.eventdrivenarchitecture.registry;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class EurekaServerApplicationTests {
+class RegistryTests {
 
 	@Test
 	void contextLoads() {
